+++ src/features/agreements-accordion/model/agreement.js
@@ -4,11 +4,19 @@ export default class Agreement {
   authority = '';
   reference = '';
   label = '';
+  resourceName = '';
 
-  constructor({ id, refId, authorityType, label }) {
+  constructor({
+    id,
+    refId,
+    authorityType,
+    label,
+    resourceName,
+  }) {
     this.id = id;
     this.reference = refId;
     this.authority = authorityType;
     this.label = label;
+    this.resourceName = resourceName;
   }
 }